requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
psycopg2-binary>=2.9.1
pyserial>=3.5
python-multipart>=0.0.5
pydantic>=1.8.2
python-dotenv>=0.19.0